+++ moveit_ros/trajectory_cache/CHANGELOG.rst
@@ -4,7 +4,8 @@ Changelog for package moveit_ros_trajectory_cache
 
 2.11.0 (2024-09-16)
 -------------------
+* Refactor `moveit_ros/trajectory_cache` package to allow feature extraction and cache insert policy injection
 
 0.1.0 (2024-05-17)
 ------------------
-* Add ``moveit_ros_trajectory_cache`` package for trajectory caching.
+* Add ``moveit_ros/trajectory_cache`` package for trajectory caching.